Issue #59 Moving locale module to top level in source as it really is a low-level module (application depends on it, not the other way around)
