+++ src/components/language_switcher/language_switcher.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { View, Text, Button } from 'react-native';
 import { Trans } from '@lingui/react';
-import { Locale } from '../../application/locale';
+import { Locale } from '../../locale';
 
 export interface Props {
     readonly currentLocale: Locale;
