+++ src/components/questionnaire/index.ts
@@ -5,7 +5,7 @@ import { Store } from '../../application/store';
 import { Id, SelectAnswerAction, selectAnswer } from '../../stores/questionnaire';
 import { selectQuestionnaire } from '../../selectors/questionnaire';
 import { selectLocale } from '../../selectors/locale';
-import { LocaleManager } from '../../application/locale';
+import { LocaleManager } from '../../locale';
 
 const mapStateToProps = (store: Store): Props => {
     const locale = LocaleManager.getLocale(selectLocale(store));
