+++ src/locale/effects.ts
@@ -1,6 +1,6 @@
 import { I18nManager, AsyncStorage } from 'react-native';
 import { Locale } from './types';
-import { PREFERENCES_LOCALE_CODE } from '../constants';
+import { PREFERENCES_LOCALE_CODE } from '../application/constants';
 
 export function isReloadNeeded(locale: Locale): boolean {
     return I18nManager.isRTL !== locale.isRTL;
