+++ src/locale/index.ts
