+++ src/locale/types.ts
